# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Price Agent Browser API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CompareRequest {
  /** Product name or URL to search for */
  query: string;
}

export interface PlatformResult {
  /** Platform name (Amazon, Flipkart, Croma) */
  platform: string;
  /** Product title */
  title?: string;
  /** Listed price in INR */
  price?: number | null;
  /** Original price before discount in INR */
  originalPrice?: number | null;
  /** Discount percentage */
  discount?: number | null;
  /** Product rating out of 5 */
  rating?: number | null;
  /** Number of reviews */
  reviewCount?: number | null;
  /** Shipping cost in INR (0 if free) */
  shippingCost?: number | null;
  /** Estimated delivery time */
  deliveryEstimate?: string | null;
  /** Effective price including shipping */
  effectivePrice?: number | null;
  /** Product URL */
  link: string;
  /** Product image URL */
  imageUrl?: string | null;
  /** Whether product is in stock */
  inStock: boolean;
  /** Error message if scraping failed */
  error?: string | null;
}

export interface CompareResponse {
  query: string;
  results: PlatformResult[];
  bestDeal?: PlatformResult | null;
  searchId?: number | null;
  timestamp: string;
}

export interface SearchRecord {
  id: number;
  query: string;
  results: PlatformResult[];
  bestDeal?: PlatformResult | null;
  timestamp: string;
}

export interface HistoryResponse {
  history: SearchRecord[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetSearchHistoryParams = {
  limit?: number;
};