# lib/api-zod/src/generated/types/compareRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Price Agent Browser API
 * OpenAPI spec version: 0.1.0
 */

export interface CompareRequest {
  /** Product name or URL to search for */
  query: string;
}