# lib/api-zod/src/generated/types/compareResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Price Agent Browser API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformResult } from "./platformResult";

export interface CompareResponse {
  query: string;
  results: PlatformResult[];
  bestDeal?: PlatformResult | null;
  searchId?: number | null;
  timestamp: Date;
}