# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Price Agent Browser API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchRecord } from "./searchRecord";

export interface HistoryResponse {
  history: SearchRecord[];
}